Add data.csv with list of authors in SPM of AR3 WGI
I copied the list of authors from the PDF in vim,
replaced commas with newlines, removed leading space
and empty lines, and copied the list to a new OpenOffice
Calc document. I then compared the name of each author
with the name displayed in the PDF, and found no error.